Reducing the amount of tasks so that it can work
The code flow through the tasks works now. The keil board can only handle 6 tasks so updateXY and update XZ were merged. Display led was merged into check luckbutton. This is probably ready for graphics.

+++ IOControls.h
@@ -3,8 +3,8 @@
 
 void initIOControls(void);
 void setLED(int);
-float checkPotentiometer(void);
-float checkJoyStick(float);
+int checkPotentiometer(void);
+int checkJoyStick(int);
 int ButtonCurrentlyPressed(void);
 
 #endif
